requirements.txt
Flask==0.12.2
line-bot-sdk==1.17.0
